move class SecureChannelSecuritySettings into a own source file
