+++ src/OpcUaStackCore/SecureChannel/SecureChannel.cpp
@@ -23,82 +23,7 @@
 
 namespace OpcUaStackCore
 {
-	// ------------------------------------------------------------------------
-	// ------------------------------------------------------------------------
-	//
-	// SecureChannelSecuritySettings
-	//
-	// ------------------------------------------------------------------------
-	// ------------------------------------------------------------------------
-	SecureChannelSecuritySettings::SecureChannelSecuritySettings(void)
-	: cryptoBase_()
-	, partnerCertificate_()
-	, clientNonce_()
-	, serverNonce_()
-	, securityKeySetClient_()
-	, securityKeySetServer_()
-	{
-	}
 
-	SecureChannelSecuritySettings::~SecureChannelSecuritySettings(void)
-	{
-	}
-
-	void
-	SecureChannelSecuritySettings::cryptoBase(CryptoBase::SPtr& cryptoBase)
-	{
-		cryptoBase_ = cryptoBase;
-	}
-
-	CryptoBase::SPtr&
-	SecureChannelSecuritySettings::cryptoBase(void)
-	{
-		return cryptoBase_;
-	}
-
-	void
-	SecureChannelSecuritySettings::partnerCertificate(Certificate::SPtr& partnerCertificate)
-	{
-		partnerCertificate_ = partnerCertificate;
-	}
-
-	Certificate::SPtr&
-	SecureChannelSecuritySettings::partnerCertificate(void)
-	{
-		return partnerCertificate_;
-	}
-
-	MemoryBuffer&
-	SecureChannelSecuritySettings::clientNonce(void)
-	{
-		return clientNonce_;
-	}
-
-	MemoryBuffer&
-	SecureChannelSecuritySettings::serverNonce(void)
-	{
-		return serverNonce_;
-	}
-
-	SecurityKeySet&
-	SecureChannelSecuritySettings::securityKeySetClient(void)
-	{
-		return securityKeySetClient_;
-	}
-
-	SecurityKeySet&
-	SecureChannelSecuritySettings::securityKeySetServer(void)
-	{
-		return securityKeySetServer_;
-	}
-
-	// ------------------------------------------------------------------------
-	// ------------------------------------------------------------------------
-	//
-	// SecureChannel
-	//
-	// ------------------------------------------------------------------------
-	// ------------------------------------------------------------------------
 	OpcUaUInt32 SecureChannel::gChannelId_ = 0;
 
 	SecureChannel::SecureChannel(IOThread* ioThread)
