+++ src/OpcUaStackCore/SecureChannel/SecureChannel.h
@@ -23,6 +23,7 @@
 #include "OpcUaStackCore/Utility/IOThread.h"
 #include "OpcUaStackCore/Certificate/CryptoBase.h"
 #include "OpcUaStackCore/Certificate/SecurityKeySet.h"
+#include "OpcUaStackCore/SecureChannel/SecureChannelSecuritySettings.h"
 #include "OpcUaStackCore/SecureChannel/MessageHeader.h"
 #include "OpcUaStackCore/SecureChannel/SecurityHeader.h"
 #include "OpcUaStackCore/SecureChannel/SecureChannelTransaction.h"
@@ -33,34 +34,6 @@
 
 namespace OpcUaStackCore
 {
-
-	class DLLEXPORT SecureChannelSecuritySettings
-	{
-	  public:
-		SecureChannelSecuritySettings(void);
-		~SecureChannelSecuritySettings(void);
-
-		void cryptoBase(CryptoBase::SPtr& cryptoBase);
-		CryptoBase::SPtr& cryptoBase(void);
-		void partnerCertificate(Certificate::SPtr& partnerCertificate);
-		Certificate::SPtr& partnerCertificate(void);
-		MemoryBuffer& clientNonce(void);
-		MemoryBuffer& serverNonce(void);
-
-		SecurityKeySet& securityKeySetClient(void);
-		SecurityKeySet& securityKeySetServer(void);
-
-	  private:
-		CryptoBase::SPtr cryptoBase_;
-		Certificate::SPtr partnerCertificate_;
-		MemoryBuffer clientNonce_;
-		MemoryBuffer serverNonce_;
-
-		SecurityKeySet securityKeySetClient_;
-		SecurityKeySet securityKeySetServer_;
-	};
-
-
 
 	class DLLEXPORT SecureChannel
 	: public TCPConnection
